requirements.txt
altgraph==0.17.4
certifi==2024.12.14
charset-normalizer==3.4.1
idna==3.10
packaging==24.2
patool==3.1.0
pefile==2023.2.7
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2025.1
pywin32-ctypes==0.2.3
requests==2.32.3
setuptools==75.1.0
urllib3==2.3.0
wheel==0.44.0
